+++ Controls-demo/progress/Rating/EmptyIconStyle/Index.ts
@@ -1,0 +1,10 @@
+import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
+import * as Template from 'wml!Controls-demo/progress/Rating/EmptyIconStyle/Template';
+
+export default class Rating extends Control<IControlOptions> {
+    protected _template: TemplateFunction = Template;
+    protected _emptyIconStyle: string[] = ['warning', 'info', 'success', 'danger', 'secondary', 'primary', 'default',
+        'contrast', 'readonly'];
+
+    static _styles: string[] = ['Controls-demo/Controls-demo'];
+}
