+++ Controls-demo/progress/Rating/EmptyIconStyle/Template.wml
@@ -1,0 +1,11 @@
+<div class="controlsDemo__wrapper controlsDemo_fixedWidth350">
+   <ws:for data="emptyIconStyle in _emptyIconStyle">
+      <div class="controlsDemo__cell">
+         <div class="controls-text-label">{{emptyIconStyle}}</div>
+         <Controls.progress:Rating readOnly="{{true}}"
+                                   emptyIconStyle="{{emptyIconStyle}}"
+                                   attr:style="{{emptyIconStyle === 'contrast' ? 'background: red;'}}"
+         />
+      </div>
+   </ws:for>
+</div>
