+++ Controls-demo/progress/Rating/IconSize/Template.wml
@@ -1,0 +1,11 @@
+<div class="controlsDemo__wrapper controlsDemo_fixedWidth350">
+   <ws:for data="iconSize in _iconSize">
+      <div class="controlsDemo__cell">
+         <div class="controls-text-label">{{iconSize}}</div>
+         <Controls.progress:Rating value="{{_value}}"
+                                   readOnly="{{true}}"
+                                   iconSize="{{iconSize}}"
+         />
+      </div>
+   </ws:for>
+</div>
