+++ Controls-demo/progress/Rating/IconStyle/Template.wml
@@ -1,0 +1,12 @@
+<div class="controlsDemo__wrapper controlsDemo_fixedWidth350">
+   <ws:for data="iconStyle in _iconStyle">
+      <div class="controlsDemo__cell">
+         <div class="controls-text-label">{{iconStyle}}</div>
+         <Controls.progress:Rating value="{{_value}}"
+                                   readOnly="{{true}}"
+                                   iconStyle="{{iconStyle}}"
+                                   attr:style="{{iconStyle === 'contrast' ? 'background: red;'}}"
+         />
+      </div>
+   </ws:for>
+</div>
