+++ Controls-demo/progress/Rating/Index.ts
@@ -1,0 +1,9 @@
+import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
+import * as Template from 'wml!Controls-demo/progress/Rating/Template';
+
+class Rating extends Control<IControlOptions> {
+    protected _template: TemplateFunction = Template;
+    static _styles: string[] = ['Controls-demo/Controls-demo'];
+}
+
+export default Rating;
