+++ Controls-demo/progress/Rating/Template.wml
@@ -1,0 +1,44 @@
+<div class="controlsDemo__wrapper">
+   <div class="controlsDemo_fixedWidth350">
+      <div class="controlsDemo__cell">
+         <Router.router:Reference attr:class="controls-text-label"
+                                  state="/Controls-demo/app/:app"
+                                  app="{{ 'Controls-demo/progress/Rating/Base/Index' }}">
+            <a href="{{content.href}}">Базовая настройка компонента рейтинга</a>
+         </Router.router:Reference>
+         <Controls-demo.progress.Rating.Base.Index/>
+      </div>
+      <div class="controlsDemo__cell">
+         <Router.router:Reference attr:class="controls-text-label"
+                                  state="/Controls-demo/app/:app"
+                                  app="{{ 'Controls-demo/progress/Rating/IconSize/Index' }}">
+            <a href="{{content.href}}">Настройка размера иконки</a>
+         </Router.router:Reference>
+         <Controls-demo.progress.Rating.IconSize.Index/>
+      </div>
+      <div class="controlsDemo__cell">
+         <Router.router:Reference attr:class="controls-text-label"
+                                  state="/Controls-demo/app/:app"
+                                  app="{{ 'Controls-demo/progress/Rating/IconPadding/Index' }}">
+            <a href="{{content.href}}">Настройка расстояния между иконками</a>
+         </Router.router:Reference>
+         <Controls-demo.progress.Rating.IconPadding.Index/>
+      </div>
+      <div class="controlsDemo__cell">
+         <Router.router:Reference attr:class="controls-text-label"
+                                  state="/Controls-demo/app/:app"
+                                  app="{{ 'Controls-demo/progress/Rating/IconStyle/Index' }}">
+            <a href="{{content.href}}">Настройка цвета иконок звезды</a>
+         </Router.router:Reference>
+         <Controls-demo.progress.Rating.IconStyle.Index/>
+      </div>
+      <div class="controlsDemo__cell">
+         <Router.router:Reference attr:class="controls-text-label"
+                                  state="/Controls-demo/app/:app"
+                                  app="{{ 'Controls-demo/progress/Rating/EmptyIconStyle/Index' }}">
+            <a href="{{content.href}}">Настройка цвета незаполненных иконок звезды</a>
+         </Router.router:Reference>
+         <Controls-demo.progress.Rating.EmptyIconStyle.Index/>
+      </div>
+   </div>
+</div>
